# src/main/resources/api-docs.yaml
openapi: 3.0.3
info:
  title: Financial Management API
  description: |
    A comprehensive REST API for financial management system that handles users, accounts, and transactions.
    
    ## Features
    - User management (CRUD operations)
    - Account management with different account types
    - Transaction processing and history
    - Real-time balance updates
    - Transaction status tracking
    
    ## Authentication
    Currently, this API does not require authentication. In production, implement proper security measures.
    
    ## Rate Limiting
    API calls are limited to 1000 requests per hour per IP address.
    
  version: 1.0.0
  contact:
    name: Financial API Support
    email: support@financial.com
    url: https://financial.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api
    description: Development server
  - url: https://api.financial.com/v1
    description: Production server
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve a list of all users in the system
      operationId: getAllUsers
      parameters:
        - name: page
          in: query
          description: Page number (0-based)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          description: Sort criteria (field,direction)
          schema:
            type: string
            example: "createdAt,desc"
      responses:
        '200':
          description: Successfully retrieved users
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pageable:
                    $ref: '#/components/schemas/Pageable'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Users
      summary: Create a new user
      description: Create a new user in the system
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request - validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '409':
          description: Conflict - user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve a specific user by their ID
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Update user
      description: Update an existing user
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request - validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete a user from the system
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    User:
      type: object
      required:
        - id
        - username
        - email
        - firstName
        - lastName
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the user
          example: 1
        username:
          type: string
          maxLength: 50
          description: Unique username
          example: "john_doe"
        email:
          type: string
          format: email
          maxLength: 100
          description: User's email address
          example: "john.doe@email.com"
        firstName:
          type: string
          maxLength: 50
          description: User's first name
          example: "John"
        lastName:
          type: string
          maxLength: 50
          description: User's last name
          example: "Doe"
        createdAt:
          type: string
          format: date-time
          description: User creation timestamp
          example: "2023-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: User last update timestamp
          example: "2023-01-01T00:00:00Z"
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - firstName
        - lastName
      properties:
        username:
          type: string
          maxLength: 50
          description: Unique username
          example: "john_doe"
        email:
          type: string
          format: email
          maxLength: 100
          description: User's email address
          example: "john.doe@email.com"
        firstName:
          type: string
          maxLength: 50
          description: User's first name
          example: "John"
        lastName:
          type: string
          maxLength: 50
          description: User's last name
          example: "Doe"
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
          description: Unique username
          example: "john_doe"
        email:
          type: string
          format: email
          maxLength: 100
          description: User's email address
          example: "john.doe@email.com"
        firstName:
          type: string
          maxLength: 50
          description: User's first name
          example: "John"
        lastName:
          type: string
          maxLength: 50
          description: User's last name
          example: "Doe"
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
          description: Page number (0-based)
          example: 0
        pageSize:
          type: integer
          description: Number of items per page
          example: 20
        sort:
          type: object
          properties:
            sorted:
              type: boolean
              example: true
            unsorted:
              type: boolean
              example: false
            empty:
              type: boolean
              example: false
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2023-01-01T00:00:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Error message
          example: "Validation failed"
        path:
          type: string
          description: Request path
          example: "/api/users"
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            validationErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: "email"
                  message:
                    type: string
                    example: "Email format is invalid"
                  rejectedValue:
                    type: string
                    example: "invalid-email"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []